+++ src/CONFIG_TYPE.mjs
@@ -1,0 +1,7 @@
+export const CONFIG_TYPE_BOOLEAN = "boolean";
+
+export const CONFIG_TYPE_NONE = "none";
+
+export const CONFIG_TYPE_NUMBER = "number";
+
+export const CONFIG_TYPE_STRING = "string";